move show contents to logical_show
